# docs/api-docs/master-data-swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Master Data",
    "description" : "Personal Master Data API.",
    "version" : "0.1.0",
    "contact" : {
      "name" : "Adam Korzeniak",
      "email" : "korzeniak.adam@gmail.com",
      "url" : "https://www.adamkorzeniak.pl"
    }
  },
  "servers" : [ {
    "description" : "Stage/Test Server",
    "url" : "https://adamkorzeniak.pl:8443/v0"
  } ],
...
  "paths" : {
    "/User/me" : {
      "get" : {
        "tags" : [ "User" ],
        "operationId" : "getMe",
        "summary" : "Retrieves details of authenticated user",
        "description" : "Retrieves details of user whose credentials were passed in request",
        "responses" : {
          "200" : {
            "$ref" : "#/components/responses/userDetails"
          },
          "400" : {
            "$ref" : "#/components/responses/badRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/unauthorized"
          },
          "404" : {
            "$ref" : "#/components/responses/notFound"
          },
          "405" : {
            "$ref" : "#/components/responses/methodNotAllowed"
          },
          "500" : {
            "$ref" : "#/components/responses/internalError"
          }
        }
      }
    },
    "/User/register" : {
      "post" : {
        "tags" : [ "User" ],
        "operationId" : "registerUser",
        "summary" : "Registers an user",
        "description" : "Registers an user with provided details",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/userRegistrationDetails"
        },
        "responses" : {
          "200" : {
            "$ref" : "#/components/responses/userRegisteredDetails"
          },
          "400" : {
            "$ref" : "#/components/responses/badRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/unauthorized"
          },
          "404" : {
            "$ref" : "#/components/responses/notFound"
          },
          "405" : {
            "$ref" : "#/components/responses/methodNotAllowed"
          },
          "409" : {
            "$ref" : "#/components/responses/conflict"
          },
          "500" : {
            "$ref" : "#/components/responses/internalError"
          }
        }
      }
    },
...
  },
  "security" : [ {
    "basicAuth" : [ ]
  } ],
  "components" : {
    "securitySchemes" : {
      "basicAuth" : {
        "type" : "http",
        "scheme" : "basic",
        "description" : "For Stage environment use below credentials:<br>Username: <b>test</b><br>Password: <b>test123</b><br><br>"
      }
    },
    "schemas" : {
      "User" : {
        "type" : "object",
        "description" : "Represents User Entity",
        "required" : [ "username", "password" ],
        "properties" : {
          "id" : {
            "description" : "User Id",
            "type" : "integer",
            "format" : "int64",
            "example" : 11,
            "minimum" : 1,
            "readOnly" : true
          },
          "username" : {
            "description" : "User username",
            "type" : "string",
            "example" : "adam",
            "minLength" : 1
          },
          "password" : {
            "description" : "User password",
            "type" : "string",
            "example" : "Pa5$w0rd",
            "minLength" : 1,
            "writeOnly" : true
          },
          "role" : {
            "description" : "User role",
            "type" : "string",
            "example" : "USER",
            "enum" : [ "USER", "ADMIN" ]
          }
        }
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "description" : "error code",
            "type" : "string",
            "example" : "REQ001"
          },
          "title" : {
            "description" : "error title",
            "type" : "string",
            "example" : "BAD REQUEST"
          },
          "message" : {
            "description" : "error message",
            "type" : "string",
            "example" : "Missing field title"
          }
        }
      }
    },
    "examples" : {
      "BadRequestErrorResponse" : {
        "value" : {
          "code" : "BAD_REQUEST",
          "title" : "Bad Request",
          "message" : "Bad Request"
        }
      },
      "NotAuthorizedErrorResponse" : {
        "value" : {
          "code" : "NOT_AUTHORIZED",
          "title" : "Not Authorized",
          "message" : "Not Authorized"
        }
      },
      "NotFoundErrorResponse" : {
        "value" : {
          "code" : "NOT_Found",
          "title" : "Not Found",
          "message" : "Not Found"
        }
      },
      "MethodNotAllowedErrorResponse" : {
        "value" : {
          "code" : "METHOD_NOT_ALLOWED",
          "title" : "Not Found",
          "message" : "Not Found"
        }
      },
      "ConflictErrorResponse" : {
        "value" : {
          "code" : "Conflict",
          "title" : "Not Found",
          "message" : "Not Found"
        }
      },
      "InternalErrorResponse" : {
        "value" : {
          "code" : "INTERNAL_ERROR",
          "title" : "Not Found",
          "message" : "Not Found"
        }
      }
    },
    "requestBodies" : {
      "userRegistrationDetails" : {
        "description" : "User details",
        "required" : true,
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/User"
            }
          }
        }
      },
...
    },
    "responses" : {
      "userDetails" : {
        "description" : "User details returned",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/User"
            }
          }
        }
      },
...
      "userRegisteredDetails" : {
        "description" : "Registered user details returned",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/User"
            }
          }
        }
      },
      "badRequest" : {
        "description" : "The request could not be understood by the server due to malformed syntax",
        "content" : {
          "application/json" : {
            "example" : {
              "$ref" : "#/components/examples/BadRequestErrorResponse"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "unauthorized" : {
        "description" : "User is not authorised",
        "content" : {
          "application/json" : {
            "example" : {
              "$ref" : "#/components/examples/NotAuthorizedErrorResponse"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "notFound" : {
        "description" : "Server could not found requested resource",
        "content" : {
          "application/json" : {
            "example" : {
              "$ref" : "#/components/examples/NotFoundErrorResponse"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "methodNotAllowed" : {
        "description" : "Method is not supported for given resource",
        "content" : {
          "application/json" : {
            "example" : {
              "$ref" : "#/components/examples/MethodNotAllowedErrorResponse"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "conflict" : {
        "description" : "Data to be created already exists.",
        "content" : {
          "application/json" : {
            "example" : {
              "$ref" : "#/components/examples/ConflictErrorResponse"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "internalError" : {
        "description" : "The server encountered an unexpected condition which prevented it from fulfilling the request.",
        "content" : {
          "application/json" : {
            "example" : {
              "$ref" : "#/components/examples/InternalErrorResponse"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}